server/requirements.txt
fastapi
uvicorn
mcp[cli]
httpx
anthropic
python-multipart
pdf2image
aiofiles
python-dotenv
rdkit
